We don't need FilePickerApiControler for two methods used only by tree, so we can rename FolderTreeControler to more consistency name FilePickerTreeController
