+++ Src/Our.Umbraco.FilePicker/App_Plugins/FilePicker/folderpickerdialog.html
@@ -1,7 +1,7 @@
 <div class="umb-panel" ng-controller="Our.Umbraco.FolderPickerDialogController">
     <div class="umb-panel-body no-header">
-        <umb-tree section="dummy"
-                  treealias="fileTree"
+        <umb-tree section="filePickerDialog"
+                  treealias="filePickerTree"
                   eventhandler="dialogEventHandler"
                   hideoptions="true"
                   hideheader="true"
